refactor: move MigrationType to own file
